requirements.txt
Faker==33.0.0
psycopg2==2.9.10
robotframework==7.1.1
robotframework-appiumlibrary==2.1.0
robotframework-assertion-engine==3.0.3
robotframework-browser==18.9.1
robotframework-jsonlibrary==0.5
robotframework-pythonlibcore==4.4.1
robotframework-requests==0.9.7
python_dotenv==1.0.1